move Reminder into Model package
